Refactor train+model impl w/ optim module
